+++ MaxMind.MinFraud/Util/DateConverter.cs
@@ -1,0 +1,14 @@
+﻿#region
+using Newtonsoft.Json.Converters;
+#endregion
+
+namespace MaxMind.MinFraud.Util
+{
+    internal class DateConverter : IsoDateTimeConverter
+    {
+        public DateConverter()
+        {
+            base.DateTimeFormat = "yyyy-MM-dd";
+        }
+    }
+}